Folder structure refactor
