+++ src/App.jsx
@@ -1,7 +1,7 @@
 import React from "react";
 import styled from "styled-components";
-import Content from "./component/Content";
-import TopBar from "./component/TopBar";
+import Home from "./component/page/Home";
+import TopBar from "./component/organism/TopBar";
 
 const AppWrapper = styled.div`
   height: 100%;
@@ -11,7 +11,7 @@ function App() {
   return (
     <AppWrapper>
       <TopBar />
-      <Content />
+      <Home />
     </AppWrapper>
   );
 }
