+++ src/component/molecule/CharacterCard/CharacterCard.jsx
