+++ src/component/molecule/CharacterCard/index.js
