+++ src/component/molecule/ComicDetail/index.js
