+++ src/component/molecule/ComicItem/ComicItem.jsx
