+++ src/component/organism/ComicList/ComicList.jsx
@@ -1,7 +1,7 @@
 import React from "react";
 import PropTypes from "prop-types";
 import styled from "styled-components";
-import ComicItem from "../ComicItem";
+import ComicItem from "../../molecule/ComicItem";
 
 const ComicListWrapper = styled.ul`
   box-sizing: border-box;
